% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{check_for_pkgdown_package}
\alias{check_for_pkgdown_package}
\title{Checks that the pkgdown package is installed.}
\usage{
check_for_pkgdown_package()
}
\description{
Checks that the pkgdown package is installed.
}
