% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{check_for_staticdocs_package}
\alias{check_for_staticdocs_package}
\title{Checks that the staticdocs package is installed.}
\usage{
check_for_staticdocs_package()
}
\description{
Checks that the staticdocs package is installed.
}

