% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile a Rocco Page From a Template.}
\usage{
compile(pkg_dir, template, out_file)
}
\arguments{
\item{pkg_dir}{character. A package directory with rocco-documented R files.}

\item{template}{character. The whisker template to use.}

\item{out_file}{character. The output file.}
}
\description{
Compile a Rocco Page From a Template.
}
