% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{create_inst}
\alias{create_inst}
\title{Create the inst directory.}
\usage{
create_inst(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Create the inst directory.
}
