% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{create_pkgdown_folder}
\alias{create_pkgdown_folder}
\title{Create the pkgdown directory.}
\usage{
create_pkgdown_folder(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Create the pkgdown directory.
}
