% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{create_staticdocs_folder}
\alias{create_staticdocs_folder}
\title{Create the staticdocs directory.}
\usage{
create_staticdocs_folder(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for staticdocs.}
}
\description{
Create the staticdocs directory.
}

