% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{create_staticdocs_index}
\alias{create_staticdocs_index}
\title{Create the staticdocs index.}
\usage{
create_staticdocs_index(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for staticdocs.}
}
\description{
Create the staticdocs index.
}

