% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{inst_exists}
\alias{inst_exists}
\title{Check whether the inst folder exists.}
\usage{
inst_exists(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Check whether the inst folder exists.
}
