% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{load_staticdocs}
\alias{load_staticdocs}
\title{Add staticdocs into the Rocco directory.}
\usage{
load_staticdocs(dir)
}
\arguments{
\item{dir}{character. The directory of the Rocco skeleton.}
}
\description{
Since Rocco and Staticdocs conflict for gh-pages and we often want both,
this will resolve the tension and create one harmonious site with rocco
docs located at index.html and staticdocs located at staticdocs/index.html.
}

