% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{pkgdown_exist}
\alias{pkgdown_exist}
\title{Check whether pkgdown exist.}
\usage{
pkgdown_exist(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Check whether pkgdown exist.
}
