% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{pkgdown_folder_exists}
\alias{pkgdown_folder_exists}
\title{Check whether the pkgdown folder exists.}
\usage{
pkgdown_folder_exists(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Check whether the pkgdown folder exists.
}
