% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{pkgdown_index_exists}
\alias{pkgdown_index_exists}
\title{Check whether a pkgdown index file exists.}
\usage{
pkgdown_index_exists(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Check whether a pkgdown index file exists.
}
