% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{staticdocs_exist}
\alias{staticdocs_exist}
\title{Check whether staticdocs exist.}
\usage{
staticdocs_exist(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for staticdocs.}
}
\description{
Check whether staticdocs exist.
}

