% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{staticdocs_folder_exists}
\alias{staticdocs_folder_exists}
\title{Check whether the staticdocs folder exists.}
\usage{
staticdocs_folder_exists(directory)
}
\description{
Check whether the staticdocs folder exists.
}

