% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{staticdocs_written}
\alias{staticdocs_written}
\title{Check whether staticdoc files have been written.}
\usage{
staticdocs_written(directory)
}
\description{
Check whether staticdoc files have been written.
}

