% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{write_pkgdown}
\alias{write_pkgdown}
\title{Writes pkgdown if they don't already exist.}
\usage{
write_pkgdown(package_dir)
}
\arguments{
\item{package_dir}{character. The directory of the package to write pkgdown for.}
}
\description{
Writes pkgdown if they don't already exist.
}
