% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{create_pkgdown_index}
\alias{create_pkgdown_index}
\title{Create the pkgdown index.}
\usage{
create_pkgdown_index(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Create the pkgdown index.
}

