% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdowns.R
\name{dir_create}
\alias{dir_create}
\title{Create a directory if it doesn't exist.}
\usage{
dir_create(directory, ...)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}

\item{...}{list. The folder structure to pass to \code{file.path}.}
}
\description{
Create a directory if it doesn't exist.
}
