% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{load_pkgdown}
\alias{load_pkgdown}
\title{Add pkgdown into the Rocco directory.}
\usage{
load_pkgdown(directory, output)
}
\arguments{
\item{directory}{character. The directory Rocco is running in.}

\item{output}{character. The directory to create the skeleton in.}
}
\description{
Since Rocco and Staticdocs conflict for gh-pages and we often want both,
this will resolve the tension and create one harmonious site with rocco
docs located at index.html and pkgdown located at pkgdown/index.html.
}

