% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{pkgdown_written}
\alias{pkgdown_written}
\title{Check whether pkgdown files have been written.}
\usage{
pkgdown_written(directory)
}
\arguments{
\item{directory}{character. The directory of the package to check for pkgdown.}
}
\description{
Check whether pkgdown files have been written.
}
