% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.rocco.R, R/rocco.R
\docType{package}
\name{rocco}
\alias{rocco}
\alias{rocco-package}
\title{Literate Documentation For R.}
\usage{
rocco(directory, output_dir = tempdir(), browse = interactive(),
  gh_pages = FALSE)
}
\arguments{
\item{directory}{character. The package directory to document.}

\item{output_dir}{character. The directory to output the static HTML site.
By default, an extemporaneously-generated temporary directory.}

\item{browse}{logical. Whether or not to launch the documentation
immediately for browsing. This will be set to \code{\link{interactive}()},
that is, TRUE if the R session is running interactive and FALSE
otherwise.}

\item{gh_pages}{logical. If set to true, rocco docs will be served on
your gh-pages branch.}
}
\value{
TRUE or FALSE according as the documentation process succeeds.
   \code{output_dir} and the launching of the browser if
   \code{browse = TRUE}.
}
\description{
Literate Documentation For R.

Given a local package directory, turn the package's R files
into a literately documented website.
}
\examples{
\dontrun{
  rocco("/path/to/package") # Will create a temporary directory and
    # display literate documentation for everything in the R directory
    # of the package at /path/to/package.

  # The below will simply create a static HTML site without opening it.
  rocco("/path/to/package", output_dir = "/my/html/dir", browse = FALSE)
}
}

