% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skeleton.R
\name{rocco_skeleton}
\alias{rocco_skeleton}
\title{Create a Skeleton for Rocco Documentation.}
\usage{
rocco_skeleton(dir)
}
\arguments{
\item{dir}{character. The directory to create the skeleton in.}
}
\description{
Given a target directory, this will copy over the necessary assets to
create the initial docco template.
}

