% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticdocs.R
\name{staticdocs_index_exists}
\alias{staticdocs_index_exists}
\title{Check whether a staticdoc index file exists.}
\usage{
staticdocs_index_exists(directory)
}
\description{
Check whether a staticdoc index file exists.
}

